refactor: decouple tools from channel, extract feishu API layer
- New feishu/api.py: standalone API functions (fetch_message_content,
  fetch_chat_history, resolve_user_name) that take a lark.Client param
- tools.py: no longer imports or references FeishuChannel; gets api_client
  from context.state['_feishu_api_client'] (same pattern as bub-schedule)
- channel.py: injects api_client into message context for tool access;
  removed all fetch/resolve methods and the _ChannelRegistry coupling
- channel.py: removed ~170 lines of API code that now lives in api.py
